+++ tests/test_models.py
@@ -3,33 +3,6 @@
 See: http://webtest.readthedocs.org/
 """
 
-import environs
-
-
-class TestConfig:
-    """Config."""
-
-    def test_config_is_set(self, testapp):
-        """Ensure required config values are set."""
-        app = testapp.app
-        assert app.config["AZURE_BASE_URL"] != "change_me"
-        assert app.config["AZURE_API_KEY"] != "change_me"
-
-    def test_default_settings_load(self, monkeypatch):
-        """Patch Env.read_env to read from .env.example and import settings."""
-        orig_read_env = environs.Env.read_env
-        monkeypatch.setattr(
-            environs.Env,
-            "read_env",
-            lambda *args, **kwargs: orig_read_env(
-                ".env.example", override=True, **kwargs
-            ),
-        )
-
-        from app import settings
-
-        assert settings.AZURE_BASE_URL == "https://change-me.openai.azure.com"
-
 
 class TestModels:
     """Models."""
